+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@ Changes:
     * library design code (detailed sample)
     * group ID (detailed sample)
     * group description (detailed sample)
+  * Added the ability to propagate library aliquots from library aliquots
   * Added readLength for libraries and library aliquots in Pinery-MISO
   * Fixed missing storage location for library aliquots in Pinery-MISO (for real this time)